devsite/requirements.txt
# Requirements needed by the devsite app server and test suite
# For initial development, we're just importing all the packages needed
# for both running the devsite server and for the pytest dependencies
#

# Versions should match those used in Open edX Ginkgo

##
## General Python package dependencies
###

celery==3.1.18

python-dateutil==2.1
path.py==8.2.1

# Yes, this is old but is the one specified by Ginkgo edx-platform
pytz==2016.7

##
## Django package dependencies
##

Django==1.8.12
django-extensions==1.5.9
djangorestframework==3.2.3
django-countries==4.0
django-filter==0.11.0
django-webpack-loader==0.5.0
django-model-utils==2.3.1

jsonfield==1.0.3  # Version used in Ginkgo. Hawthorn uses version 2.0.2

##
## Documentation (Sphinx) dependencies
##

Sphinx==1.8.1
recommonmark==0.4.0

##
## Open edX package dependencies
##

edx-opaque-keys==0.4
#edx-drf-extensions==1.2.2

##
## Multi-tenancy
##

git+https://github.com/appsembler/edx-organizations.git@0.4.1-appsembler10

##
## Pytest dependencies
##

coverage==4.5.1
factory-boy==2.5.1
pylint==1.8.2
pylint-django==0.9.1
pytest==3.1.3
pytest-django==3.1.2
pytest-mock==1.7.1
pytest-pythonpath==0.7.2
pytest-cov==2.6.0
